requirements.txt
matplotlib
tensorflow
opencv-python
scikit-image
pandas
Pillow
pysimplegui
pyimagej
scikit-learn
scipy
dash
dash-bootstrap-components